% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census70ibge}
\alias{census70ibge}
\title{IBGE's 1970 demographic census aggregated by municipality.}
\format{
O formato é:
\describe{
\item{coluna}{descricao}
...
}
}
\source{
\url{https://www.kaggle.com/shivamb/netflix-shows?select=netflix_titles.csv}
}
\usage{
census70ibge
}
\description{
This dataset contains descriptive statistics for the Brazilian municipalities calculated from the
IBGE's 1970 demographic census microdata.
}
\keyword{datasets}
