% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_censo70.R
\name{import_censo70}
\alias{import_censo70}
\title{To import data from the 70s Census from the data-raw folder}
\usage{
import_censo70(file)
}
\arguments{
\item{file}{The Census70 file name (in .csv) for a certain state}
}
\value{
The correspondent Census70 file in data.frame format
}
\description{
To import the data from the IBGE's 1970s Census stored in the data-raw folder
}
